feat: standardize DeviceDto property naming to camelCase; update related components and services for consistency

+++ sarah.client/src/app/services/api-client/api/devices.service.ts
@@ -199,7 +199,7 @@ export class DevicesService {
             }
         }
 
-        let localVarPath = `/Devices/${this.configuration.encodeParam({name: "id", value: id, in: "path", style: "simple", explode: false, dataType: "number", dataFormat: "int32"})}`;
+        let localVarPath = `/Devices/${this.configuration.encodeParam({name: "id", value: id, in: "path", style: "simple", explode: false, dataType: "number", dataFormat: "int64"})}`;
         return this.httpClient.request<any>('delete', `${this.configuration.basePath}${localVarPath}`,
             {
                 context: localVarHttpContext,
